Switch UNSIGNED_LONG to NUMERIC xDBC type
Most client applications don't support/check unsigned types.
Furthermore, seeing a BIGINT type, they assume it's signed (although
xDBC APIs allow checking). Consequently the NUMERIC type offers more conversion
safty.

+++ x-pack/plugin/sql/jdbc/src/main/java/org/elasticsearch/xpack/sql/jdbc/EsType.java
@@ -19,7 +19,7 @@ public enum EsType implements SQLType {
     SHORT(                    Types.SMALLINT),
     INTEGER(                  Types.INTEGER),
     LONG(                     Types.BIGINT),
-    UNSIGNED_LONG(            Types.BIGINT),
+    UNSIGNED_LONG(            Types.NUMERIC),
     DOUBLE(                   Types.DOUBLE),
     FLOAT(                    Types.REAL),
     HALF_FLOAT(               Types.FLOAT),
